Clean and fix some vimrc things
Fixes the rubocop config for syntastic.

Clean up some plugins, dependencies, and reorganize the config blocks to match the order of the plugins. I added SuperTab as an "eh, it's better than nothing" default in the event that the other completion plugins won't work.

CtrlP now uses c-t as the hotkey.